chore: refactor sql runner and semantic viewer shared componentes
