+++ packages/frontend/src/components/DataViz/Results.tsx
@@ -1,0 +1,6 @@
+import { type ResultRow, type SqlColumn } from '@lightdash/common';
+
+export type ResultsAndColumns = {
+    results: ResultRow[];
+    columns: SqlColumn[];
+};
