--- packages/frontend/src/features/semanticViewer/components/CartesianChartConfiguration.tsx
@@ -1,76 +1,0 @@
-import { ChartKind } from '@lightdash/common';
-import { ActionIcon, Divider, Group, Stack, Title } from '@mantine/core';
-import { IconChevronDown, IconChevronRight } from '@tabler/icons-react';
-import { useState } from 'react';
-import MantineIcon from '../../../components/common/MantineIcon';
-import { barChartConfigSlice } from '../store/barChartSlice';
-import { useAppSelector } from '../store/hooks';
-import { lineChartConfigSlice } from '../store/lineChartSlice';
-import { CartesianChartFieldConfiguration } from './CartesianChartFieldConfiguration';
-import { CartesianChartStyling } from './CartesianChartStyling';
-
-export const CartesianChartConfig = () => {
-    const [isFieldConfigurationOpen, setIsFieldConfigurationOpen] =
-        useState(true);
-    const [isStylingOpen, setIsStylingOpen] = useState(true);
-    const selectedChartType = useAppSelector(
-        (state) => state.semanticViewer.selectedChartType,
-    );
-
-    const actions =
-        selectedChartType === ChartKind.LINE
-            ? lineChartConfigSlice.actions
-            : selectedChartType === ChartKind.VERTICAL_BAR
-            ? barChartConfigSlice.actions
-            : null;
-
-    if (!actions) {
-        return null;
-    }
-
-    return (
-        <Stack spacing="xs" mb="lg">
-            <Group spacing="xs">
-                <Title order={5} fz="sm" c="gray.9">
-                    Data
-                </Title>
-                <ActionIcon>
-                    <MantineIcon
-                        onClick={() =>
-                            setIsFieldConfigurationOpen(
-                                !isFieldConfigurationOpen,
-                            )
-                        }
-                        icon={
-                            isFieldConfigurationOpen
-                                ? IconChevronDown
-                                : IconChevronRight
-                        }
-                    />
-                </ActionIcon>
-            </Group>
-
-            {isFieldConfigurationOpen && (
-                <CartesianChartFieldConfiguration actions={actions} />
-            )}
-
-            <Divider my="sm" />
-
-            <Group spacing="xs">
-                <Title order={5} fz="sm" c="gray.9">
-                    Styling
-                </Title>
-                <ActionIcon>
-                    <MantineIcon
-                        onClick={() => setIsStylingOpen(!isStylingOpen)}
-                        icon={
-                            isStylingOpen ? IconChevronDown : IconChevronRight
-                        }
-                    />
-                </ActionIcon>
-            </Group>
-
-            {isStylingOpen && <CartesianChartStyling actions={actions} />}
-        </Stack>
-    );
-};
