+++ packages/frontend/src/features/semanticViewer/store/selectors.ts
@@ -50,67 +50,11 @@ const selectChartConfigByKind = createSelector(
 
 export const selectCurrentChartConfig = createSelector(
     [selectSemanticViewerRunnerState, (state) => state],
-    (sqlRunnerState, state) => {
-        const { selectedChartType } = sqlRunnerState;
+    (semanticViewerState, state) => {
+        const { selectedChartType } = semanticViewerState;
         return selectChartConfigByKind(state, selectedChartType);
     },
 );
-
-export const selectCurrentCartesianChartState = createSelector(
-    [
-        selectSemanticViewerRunnerState,
-        selectBarChartConfigState,
-        selectLineChartConfigState,
-    ],
-    (sqlRunnerState, barChartConfig, lineChartConfig) => {
-        const { selectedChartType } = sqlRunnerState;
-        if (selectedChartType === ChartKind.VERTICAL_BAR) {
-            return barChartConfig;
-        } else if (selectedChartType === ChartKind.LINE) {
-            return lineChartConfig;
-        }
-        return null;
-    },
-);
-
-const getXLayoutOptions = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.xLayoutOptions,
-);
-
-const getYLayoutOptions = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.yLayoutOptions,
-);
-
-const getXAxisField = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.config?.fieldConfig?.x,
-);
-
-const getYAxisFields = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.config?.fieldConfig?.y,
-);
-
-const getGroupByField = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.config?.fieldConfig?.groupBy?.[0],
-);
-
-const getGroupByLayoutOptions = createSelector(
-    [selectCurrentCartesianChartState],
-    (chartConfig) => chartConfig?.options?.groupByOptions,
-);
-
-export const cartesianChartSelectors = {
-    getXLayoutOptions,
-    getYLayoutOptions,
-    getXAxisField,
-    getYAxisFields,
-    getGroupByField,
-    getGroupByLayoutOptions,
-};
 
 const selectSelectedDimensions = (state: RootState) =>
     state.semanticViewer.selectedDimensions;
