+++ packages/frontend/src/features/semanticViewer/store/tableVisSlice.ts
@@ -91,6 +91,3 @@ export const tableVisSlice = createSlice({
         });
     },
 });
-
-export const { updateFieldLabel, updateColumnVisibility } =
-    tableVisSlice.actions;
