+++ packages/frontend/src/features/sqlRunner/components/ContentPanel.tsx
@@ -24,6 +24,8 @@ import {
 import { ResizableBox } from 'react-resizable';
 import { ConditionalVisibility } from '../../../components/common/ConditionalVisibility';
 import MantineIcon from '../../../components/common/MantineIcon';
+import SqlRunnerChart from '../../../components/DataViz/visualizations/ChartView';
+import { Table } from '../../../components/DataViz/visualizations/Table';
 import RunSqlQueryButton from '../../../components/SqlRunner/RunSqlQueryButton';
 import { useSqlQueryRun } from '../hooks/useSqlQueryRun';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
@@ -36,8 +38,6 @@ import {
     setSqlRunnerResults,
 } from '../store/sqlRunnerSlice';
 import { SqlEditor } from './SqlEditor';
-import SqlRunnerChart from './visualizations/SqlRunnerChart';
-import { Table } from './visualizations/Table';
 
 const MIN_RESULTS_HEIGHT = 10;
