+++ packages/frontend/src/features/sqlRunner/components/FieldReferenceSelect.tsx
@@ -1,7 +1,7 @@
 import { type DimensionType } from '@lightdash/common';
 import { Select, type SelectProps } from '@mantine/core';
 import { type FC } from 'react';
-import { TableFieldIcon } from './TableFields';
+import { TableFieldIcon } from '../../../components/DataViz/Icons';
 
 type Props = SelectProps & {
     fieldType: DimensionType;
