+++ packages/frontend/src/features/sqlRunner/store/pieChartSlice.ts
@@ -75,5 +75,3 @@ export const pieChartConfigSlice = createSlice({
         });
     },
 });
-
-export const { setGroupFieldIds } = pieChartConfigSlice.actions;
