+++ packages/frontend/src/pages/ViewSqlChart.tsx
@@ -20,9 +20,9 @@ import { useUnmount } from 'react-use';
 import ErrorState from '../components/common/ErrorState';
 import MantineIcon from '../components/common/MantineIcon';
 import Page from '../components/common/Page/Page';
+import SqlRunnerChart from '../components/DataViz/visualizations/ChartView';
+import { Table } from '../components/DataViz/visualizations/Table';
 import { Header } from '../features/sqlRunner/components/Header';
-import SqlRunnerChart from '../features/sqlRunner/components/visualizations/SqlRunnerChart';
-import { Table } from '../features/sqlRunner/components/visualizations/Table';
 import { useSavedSqlChart } from '../features/sqlRunner/hooks/useSavedSqlCharts';
 import { useSqlChartResults } from '../features/sqlRunner/hooks/useSqlChartResults';
 import { store } from '../features/sqlRunner/store';
